+++ turbo/apps/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [9.348.1](https://github.com/okou-ai/okou/compare/cli-v9.348.0...cli-v9.348.1) (2026-09-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @okouai/api-contracts bumped to 1.478.1
+    * @okouai/core bumped to 8.687.1
+    * @okouai/pi-agent-runtime bumped to 1.33.28
+
 ## [9.348.0](https://github.com/okou-ai/okou/compare/cli-v9.347.0...cli-v9.348.0) (2026-09-20)